requirements.txt
anthropic>=0.15.0
voyageai>=0.1.5
requests>=2.28.0
beautifulsoup4>=4.12.0
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0
cohere>=4.0.0
elasticsearch>=8.0.0
weaviate-client>=3.20.0,<4.0.0
docker>=6.0.0
flask>=2.3.0
flask-socketio>=5.3.0
python-dotenv>=1.0.0